# [ENHANCEMENT] PromQL: use Kahan summation for sum()