@bboreham 2024-06-24
Current feeling is the slow-down will not be noticeable compared to overheads such as disk IO.